# openapi.yml
openapi: 3.0.3
info:
  title: Ledger V3 POC API
  description: API for the Ledger V3 Proof of Concept with Raft consensus
  version: 1.0.0
  contact:
    name: Formance
    url: https://formance.com
servers:
  - url: http://localhost:9000
    description: Local development server
security:
  - BearerAuth: []
  - {}
paths:
...
  /clusterz:
    get:
      summary: Cluster-availability probe
      description: |
        Returns 200 only when the node is part of a healthy cluster (Raft state
        healthy and a leader elected). Disk and clock skew gate writes (see
        admission), not cluster readiness. Returns 503 otherwise. Use this
        for external monitoring or any client that needs to wait until the node
        can actually serve cluster-dependent traffic. Kubernetes readiness
        probes should target `/readyz` instead — this endpoint MUST NOT gate
        StatefulSet ordering, otherwise the cluster cannot recover from a
        full-kill cold start.
      operationId: clusterz
      tags:
        - Health
      responses:
        '200':
          description: Node is part of a healthy cluster
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
  /v3/{ledgerName}/stats:
    get:
      summary: Get ledger statistics
      description: Returns aggregate statistics (account count, transaction count) for a ledger. Requires `ledger:read` scope.
      operationId: getLedgerStats
      tags:
        - Ledgers
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/LedgerName'
      responses:
        '200':
          description: Ledger statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetLedgerStatsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /v3/{ledgerName}/logs:
    get:
      summary: List logs for a ledger
      description: |
        Lists log entries for a specific ledger, ordered by ledger-local log ID ascending.
        Requires the builtin `LOG` index to be enabled on the ledger via `CreateIndex`.
        Supports cursor pagination via the `after` query parameter. Requires `ops:read` scope.
      operationId: listLedgerLogs
      tags:
        - Logs
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/LedgerName'
        - name: pageSize
          in: query
          schema:
            type: integer
            format: uint32
            minimum: 0
            default: 0
          description: Maximum number of log entries to return (0 = no limit)
        - name: after
          in: query
          schema:
            type: integer
            format: uint64
            minimum: 0
          description: Ledger-local log ID to start after (exclusive, for cursor pagination)
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
          description: |
            Filter logs with date >= startDate (inclusive, RFC3339 format).
            Requires the builtin `LOG_DATE` index to be enabled on the ledger.
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
          description: |
            Filter logs with date < endDate (exclusive, RFC3339 format).
            Requires the builtin `LOG_DATE` index to be enabled on the ledger.
      responses:
        '200':
          description: Logs retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  description: System log entry (protobuf JSON representation)
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /v3/{ledgerName}/promote:
    post:
      summary: Promote a mirror ledger
      description: |
        Promotes a mirror-mode ledger to normal mode. After promotion, the mirror sync
        worker is stopped and the ledger accepts direct write operations. Requires `ledger:write` scope.
      operationId: promoteLedger
      tags:
        - Ledgers
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/LedgerName'
      responses:
        '201':
          description: Ledger promoted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromoteLedgerResponse'
        '400':
          description: Ledger is not in mirror mode
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /v3/{ledgerName}/transactions:
    post:
      summary: Create a new transaction
      description: Creates a new transaction in the specified ledger with the specified postings. Requires `ledger:write` scope.
      operationId: createTransaction
      tags:
        - Transactions
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/LedgerName'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Transaction created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTransactionResponse'
          links:
            GetCreatedTransaction:
              operationId: getTransaction
              parameters:
                ledgerName: '$request.path.ledgerName'
                transactionId: '$response.body#/data/transaction/id'
            RevertCreatedTransaction:
              operationId: revertTransaction
              parameters:
                ledgerName: '$request.path.ledgerName'
                transactionId: '$response.body#/data/transaction/id'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflict (e.g. concurrent modification)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /v3/{ledgerName}/transactions/{transactionId}:
    get:
      summary: Get a transaction
      description: Retrieves a transaction by its ID from the specified ledger. Requires `ledger:read` scope.
      operationId: getTransaction
      tags:
        - Transactions
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/LedgerName'
        - $ref: '#/components/parameters/TransactionId'
      responses:
        '200':
          description: Transaction retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetTransactionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /v3/{ledgerName}/accounts:
    get:
      summary: List accounts
      description: Lists accounts in a ledger in alphabetical order with optional prefix filtering and cursor pagination. Requires `ledger:read` scope.
      operationId: listAccounts
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/LedgerName'
        - name: pageSize
          in: query
          description: Maximum number of accounts to return (0 = no limit)
          schema:
            type: integer
            format: int32
        - name: after
          in: query
          description: Account address to start after (for cursor pagination, exclusive)
          schema:
            type: string
        - name: prefix
          in: query
          description: Filter accounts by address prefix (e.g. "users:")
          schema:
            type: string
      responses:
        '200':
          description: List of accounts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /v3/{ledgerName}/accounts/{address}:
    get:
      summary: Get an account
      description: Retrieves account information including metadata for a specific account address. Requires `ledger:read` scope.
      operationId: getAccount
      tags:
        - Accounts
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/LedgerName'
        - $ref: '#/components/parameters/AccountAddress'
      responses:
        '200':
          description: Account retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAccountResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /v3/{ledgerName}/volumes:
    get:
      summary: Aggregate volumes
      description: |
        Returns aggregated volumes (input/output/balance per asset) across all accounts in the ledger.
        Supports filtering by account address prefix, grouping by account prefixes, and max-precision merging.
        Requires `accounts:read` scope.
      operationId: aggregateVolumes
      tags:
        - Volumes
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/LedgerName'
        - name: prefix
          in: query
          description: Filter accounts by address prefix
          schema:
            type: string
        - name: useMaxPrecision
          in: query
          description: Merge assets with the same base at the highest observed precision
          schema:
            type: boolean
            default: false
        - name: groupByPrefixes
          in: query
          description: Comma-separated list of account prefixes to group volumes by
          schema:
            type: string
      responses:
        '200':
          description: Aggregated volumes retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AggregateVolumesResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /v3/{ledgerName}/metadata:
    post:
      summary: Save ledger metadata
      description: Saves metadata key-value pairs on the ledger itself. Requires `ledger:write` scope.
      operationId: saveLedgerMetadata
      tags:
        - Ledgers
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/LedgerName'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/MetadataValue'
              description: Metadata key-value pairs (supports typed values)
              example:
                environment: production
                team: payments
                region: eu-west-1
      responses:
        '204':
          description: Ledger metadata saved successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /v3/{ledgerName}/metadata/{key}:
    delete:
      summary: Delete ledger metadata
      description: Deletes a metadata key from the ledger. Requires `ledger:write` scope.
      operationId: deleteLedgerMetadata
      tags:
        - Ledgers
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/LedgerName'
        - $ref: '#/components/parameters/MetadataKey'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204':
          description: Ledger metadata deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /v3/{ledgerName}/accounts/{address}/metadata:
    post:
      summary: Save account metadata
      description: Saves metadata for a specific account in the ledger. Requires `ledger:write` scope.
      operationId: saveAccountMetadata
      tags:
        - Accounts
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/LedgerName'
        - $ref: '#/components/parameters/AccountAddress'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/MetadataValue'
              description: Metadata key-value pairs (supports typed values)
              example:
                account_type: asset
                label: Bank Account
                active: true
                balance_threshold: 1000
      responses:
        '204':
          description: Account metadata saved successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /v3/{ledgerName}/transactions/{transactionId}/metadata:
    post:
      summary: Save transaction metadata
      description: Saves metadata for a specific transaction in the ledger. Requires `ledger:write` scope.
      operationId: saveTransactionMetadata
      tags:
        - Transactions
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/LedgerName'
        - $ref: '#/components/parameters/TransactionId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties:
                $ref: '#/components/schemas/MetadataValue'
              description: Metadata key-value pairs (supports typed values)
              example:
                reason: correction
                source: support
                priority: 1
      responses:
        '204':
          description: Transaction metadata saved successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /v3/{ledgerName}/accounts/{address}/metadata/{key}:
    delete:
      summary: Delete account metadata
      description: Deletes a metadata key for a specific account in the ledger. Requires `ledger:write` scope.
      operationId: deleteAccountMetadata
      tags:
        - Accounts
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/LedgerName'
        - $ref: '#/components/parameters/AccountAddress'
        - $ref: '#/components/parameters/MetadataKey'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204':
          description: Account metadata deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
  /v3/{ledgerName}/transactions/{transactionId}/metadata/{key}:
    delete:
      summary: Delete transaction metadata
      description: Deletes a metadata key for a specific transaction in the ledger. Requires `ledger:write` scope.
      operationId: deleteTransactionMetadata
      tags:
        - Transactions
      security:
        - BearerAuth: []
        - {}
      parameters:
        - $ref: '#/components/parameters/LedgerName'
        - $ref: '#/components/parameters/TransactionId'
        - $ref: '#/components/parameters/MetadataKey'
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204':
          description: Transaction metadata deleted successfully
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT token issued by a trusted OIDC provider. Required when the server
        is started with `--auth-enabled`. Scopes: `ledger:read` (read operations),
        `ledger:write` (write operations), `ledger:admin` (cluster operations).
  parameters:
    LedgerName:
      name: ledgerName
      in: path
      required: true
      schema:
        type: string
      description: Name of the ledger
...
  responses:
    ServiceUnavailable:
      description: Service unavailable (no leader available)
      headers:
        Retry-After:
          description: Number of seconds to wait before retrying the request
          schema:
            type: integer
            format: int32
            minimum: 1
            example: 1
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Unauthorized:
      description: Authentication required (missing or invalid JWT token)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    GetLedgerStatsResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: '#/components/schemas/LedgerStats'
...
    HealthResponse:
      type: object
      required:
        - data
      properties:
        data:
          type: object
          required:
            - status
          properties:
            status:
              type: string
              description: Health status (e.g. "ok")
...